Move protocols and ioList to utils.py
